requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
SQLAlchemy
Werkzeug
Jinja2
tensorflow-macos
tensorflow-metal
tensorflow
opencv-python
numpy
matplotlib
scikit-learn
Pillow
requests
openai
python-dotenv
gunicorn
pytest
Faker
